fixing

+++ src/services/postgres/PlaylistSongsService.js
@@ -6,23 +6,15 @@ const AuthorizationError = require("../../exceptions/AuthorizationError");
 const { mapPlaylistToModel } = require("../../utils");
 
 class PlaylistSongsService {
-  constructor() {
+  constructor(collaborationService) {
     this._pool = new Pool();
+    this._collaborationService = collaborationService;
   }
 
-  async addPlaylistSong({ playlistId, songId, owner }) {
+  async addPlaylistSong({ playlistId, songId, credentialId }) {
     const id = `playlistsong-${nanoid(16)}`;
 
-    const queryCheck = {
-      text: "SELECT * FROM playlists WHERE id = $1 AND owner = $2",
-      values: [playlistId, owner],
-    };
-
-    const resultCheck = await this._pool.query(queryCheck);
-
-    if (!resultCheck.rows.length) {
-      throw new AuthorizationError("Tidak memiliki akses ke playlist");
-    }
+    await this.verifyPlaylistAccess(playlistId, credentialId);
 
     const query = {
       text: "INSERT INTO playlistsongs VALUES($1, $2, $3) RETURNING id",
@@ -37,15 +29,17 @@ class PlaylistSongsService {
     return result.rows[0].id;
   }
 
-  async getPlaylistSongs() {
+  async getPlaylistSongs({ playlistId, credentialId }) {
+    await this.verifyPlaylistAccess(playlistId, credentialId);
     const result = await this._pool.query({
-      text: "SELECT id, playlist_id, song_id FROM playlistsongs",
-      values: [],
+      text: "SELECT id, playlist_id, song_id FROM playlistsongs WHERE playlist_id = $1",
+      values: [playlistId],
     });
     return result.rows.map(mapPlaylistToModel);
   }
 
-  async deletePlaylistSongById({ playlistId, songId }) {
+  async deletePlaylistSongById({ playlistId, songId, credentialId }) {
+    await this.verifyPlaylistAccess(playlistId, credentialId);
     const query = {
       text: "DELETE FROM playlistsongs WHERE playlist_id = $1 AND song_id = $2 RETURNING id",
       values: [playlistId, songId],
@@ -55,6 +49,40 @@ class PlaylistSongsService {
 
     if (!result.rows.length) {
       throw new NotFoundError("Playlistsong gagal dihapus. Id tidak ditemukan");
+    }
+  }
+
+  async verifyPlaylistOwner(id, owner) {
+    const query = {
+      text: "SELECT * FROM playlists WHERE id = $1",
+      values: [id],
+    };
+
+    const result = await this._pool.query(query);
+
+    if (!result.rows.length) {
+      throw new NotFoundError("Resource yang Anda minta tidak ditemukan");
+    }
+
+    const playlist = result.rows[0];
+
+    if (playlist.owner !== owner) {
+      throw new AuthorizationError("Anda tidak berhak mengakses resource ini");
+    }
+  }
+
+  async verifyPlaylistAccess(playlistId, userId) {
+    try {
+      await this.verifyPlaylistOwner(playlistId, userId);
+    } catch (error) {
+      if (error instanceof NotFoundError) {
+        throw error;
+      }
+      try {
+        await this._collaborationService.verifyCollaborator(playlistId, userId);
+      } catch {
+        throw error;
+      }
     }
   }
 }
